ex-example.sh
The is an example of using the little known EX command in linux, specifically for removing strings in a file much like you would in VI/VIM, but without having to open the file in a text editor. "Sed -i" is great, but sometime you don't want to stream or create duplicated files.